changed to CommonJS
