+++ app.js
@@ -1,26 +1,26 @@
-import express from 'express';
-import morgan from 'morgan';
+const express = require('express');
+const morgan = require('morgan');
 
-import shipmentRouter from './routes/shipmentRoutes';
-import userRouter from './routes/userRoutes';
+const shipmentRouter = require('./routes/shipmentRoutes');
+const userRouter = require('./routes/userRoutes');
 
 const app = express();
 
 // 1) Middlewares
 if (process.env.NODE_ENV === 'development') {
-    app.use(morgan('dev'));
+  app.use(morgan('dev'));
 }
 app.use(express.json());
 app.use(express.static(`${__dirname}/public`));
 
 app.use((req, res, next) => {
-    console.log('Hello from the middleware');
-    next();
+  console.log('Hello from the middleware');
+  next();
 });
 
 app.use((req, res, next) => {
-    req.requestTime = new Date().toISOString();
-    next();
+  req.requestTime = new Date().toISOString();
+  next();
 });
 
 // Routes Mounting
