+++ config.env
@@ -1,4 +1,5 @@
 NODE_ENV=development
 PORT=3000
-DATABASE=mongodb+srv://cla:<PASSWORD>@cluster0.ry4wn3c.mongodb.net/mubeatrack?retryWrites=true&w=majority
-DATABASE_PASSWORD=XFYictoPYmV7slZh
+DATABASE=mongodb+srv://mubeatrack:<password>@mubeatrack.f4asqw2.mongodb.net/mubeatrack?retryWrites=true&w=majority
+DATABASE_PASSWORD=bIy5nzkQn8tBaBz6
+
