+++ package.json
@@ -2,11 +2,11 @@
   "name": "mvc_mubeatrack",
   "version": "1.0.0",
   "description": "MubeaTrack with Express and MongoDB",
-  "main": "app.mjs",
+  "main": "app.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon server.mjs",
-    "start:prod": "SET NODE_ENV=development&&nodemon server.mjs"
+    "start": "nodemon server.js",
+    "start:prod": "SET NODE_ENV=development&&nodemon server.js"
   },
   "author": "Cla Tschenett, John Schmid",
   "license": "ISC",
