+++ routes/userRoutes.js
@@ -1,5 +1,5 @@
 const express = require("express");
-const userController = require('./../controllers/userController');
+const userController = require('../controllers/userController');
 
 
 const router = express.Router();
